Add unit tests for gradFilter
Add gradFilter unit tests which compare the SIMD and C implementations.
To make this possible, add an enableOpt parameter to
vvdec::InterPrediction::init() that toggles SIMD optimizations, and
change BioGradFilter and profGradFilter from protected to public so
the tests can call them directly. The init()'s new parameter has a
default value, so call sites do not have to be updated.