Suppress non-trivial memcall warnings in Clang and GCC
Some calls operate on pointers to non-trivially copyable types, which
triggers `-Wnontrivial-memcall` in Clang and `-Wclass-memaccess` in GCC.

To intentionally perform byte-wise initialization/copy in these sites,
cast the destination pointer to `void*` to make that intent explicit and
silence both warnings.

This was previously handled by the `GCC_WARNING_DISABLE_class_memaccess`
macro for GCC, however this does not work for Clang. With the addition
of the cast this is now unnecessary and is removed.