Arm: Port SVE implementations of filterWxH_N8 from VVenC
Port the SVE implementations for the 8-tap filterWxH_N8 from VVenC and
add the corresponding init function.

The 8-tap filter for filter4x4 is reduced to a 6-tap filter due to the
zero coefficients at the 0th and 7th indices.

These SVE implementations are around 24% faster than their Neon
counterparts when benchmarked on a Neoverse V2 using LLVM 21.